tools/z-image-turbo/requirements.txt
stable-diffusion-cpp-python>=0.5.0
pillow>=10.0.0
